put NIC unit tests into separate file
